Added Modules Configuration Dialogue
